reformatted code and modified add to cart.
